Creating a webrtc library PeerConnection object
Summary:
In order to create a webrtc PeerConnection, we must provide a
webrtc::PeerConnectionObserver object to the PeerConnection's creation method.
This observer will be the bridge between the webrtc PeerConnection object events
and WebKit.
This commit also adds a header with some definitions that are needed in order to
use the webrtc headers (since they do not provide clean headers)
We also needed to disable the DisallowCType and let Nix use ctype functions,
because webrtc library uses them.

Reviewers: lmoura, Tullio, luck, hugopl, dakerfp, setanta, sergio, nick

Differential Revision: http://phabricator.coderev.indt.org/D2919