Improve MediaStream test.
When running under certain conditions the
fake mic data sent in sourceDataVector can be ignored by
AudioDestinationNode when it's not initialized yet, e.g.
the HRTF database is still loading. This behavior, while
acceptable when using real mic data, leads to a faulty
pattern from audioDataVector being checked for in the test.

This patch improves the test by sending a series of '1' to
signal when the test pattern (0-1023) should be sent.

Reviewed by: Tullio Lucena