Fix SessionState restore and move to Nix file.
Session state restoring was failing in debug mode as the
BackForward items restored in the UIProcess were using the same
itemID as the original ones from WebProcess.

This patch fixes this behavior by using even itemIDs for UIProcess
items.

Also moved the function implementation to WebPageProxyNix to avoid
affecting other ports and easing the upstream process.

Reviewed by: Hugo Parente Lima