Removing intrusive Nix code from HistoryController
In the past we were fixing the scale and content position issue, when going back/forward to/from a page
with a patch in HistoryController.
That patch was because FrameView::setWasScrolledByUser was being called between a Frame::reset and a
HistoryController::restoreScrollPositionAndViewState, leading to the item that was being loaded to do not
have its scale and content position set properly.
That call was happening because the new FrameView (to load the new page) was being created with the old
fixedLayoutSize (from the previous page). To simply overcome that, we now create the new FrameView with
a default fixedLayoutSize.

Reviewed by: Tullio Lucena